Move DistinctLimitNode to spi
