+++ presto-main/src/main/java/com/facebook/presto/sql/planner/plan/InternalPlanVisitor.java
@@ -29,11 +29,6 @@ public abstract class InternalPlanVisitor<R, C>
         return visitPlan(node, context);
     }
 
-    public R visitDistinctLimit(DistinctLimitNode node, C context)
-    {
-        return visitPlan(node, context);
-    }
-
     public R visitSample(SampleNode node, C context)
     {
         return visitPlan(node, context);
