+++ presto-spi/src/main/java/com/facebook/presto/spi/plan/PlanVisitor.java
@@ -74,4 +74,9 @@ public abstract class PlanVisitor<R, C>
     {
         return visitPlan(node, context);
     }
+
+    public R visitDistinctLimit(DistinctLimitNode node, C context)
+    {
+        return visitPlan(node, context);
+    }
 }
